refactor: remove cycles from Common components; cycles are now an error (#257)
