+++ src/components/Common/DocumentViewer/DocumentViewer.tsx
@@ -1,8 +1,8 @@
 import { useRef } from 'react'
 import { Link } from 'react-aria-components'
+import { Flex } from '../Flex/Flex'
 import styles from './DocumentViewer.module.scss'
 import { useContainerBreakpoints } from '@/hooks/useContainerBreakpoints/useContainerBreakpoints'
-import { Flex } from '@/components/Common'
 
 interface DocumentViewerProps {
   url?: string | null
