+++ src/components/Common/PaginationControl/PaginationControl.tsx
@@ -1,7 +1,9 @@
 import { useTranslation } from 'react-i18next'
 import { useForm } from 'react-hook-form'
+import { Flex } from '../Flex/Flex'
+import { SelectField } from '../Fields/SelectField/SelectField'
+import { Button } from '../Button/Button'
 import style from './PaginationControl.module.scss'
-import { Button, Flex, SelectField } from '@/components/Common'
 import PaginationFirstIcon from '@/assets/icons/pagination_first.svg?react'
 import PaginationPrevIcon from '@/assets/icons/pagination_previous.svg?react'
 import PaginationNextIcon from '@/assets/icons/pagination_next.svg?react'
